% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade.R
\name{upgrade}
\alias{upgrade}
\title{Upgrade all installed packages}
\usage{
upgrade(daily = FALSE)
}
\arguments{
\item{daily}{If \code{TRUE}, exit early if \code{upgrade()} has already been run today}
}
\value{
Invisibly returns a list of updated packages
}
\description{
Upgrade all installed packages
}
